+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2021/04/11 [LATEST]
+## [1.2.1] - 2021/04/14 [LATEST]
+
+### Added
+
+- mention as prefix
+- Image manupulation with errors
+
+### Updated
+
+- Hidden spotify keys in music handler
+
+## [1.2.0] - 2021/04/14
 
 ### Added